Move guestfs to pkg/diskimage
This is needed to make it possible to import pkg/utils from the
flexvolume driver. Without this change, flexvolume fails trying
to load libguestfs shared library.
