+++ pkg/diskimage/guestfs/guestfs.go
